Fix SharedObject cookie errors
For the third time. This time it actually works because... get ready for it... we actually debugged it!!

This also removes the warnings and applies the settings for the current instance of the client. It will go away when the user refreshes, but that will still allow them to use it temporarily.