Rename, kongAPI edits
I was wrong about the functions in the Main class that started with the names method_689 and method_806; it was to get the Kong API when testing mode is enabled versus the normal method of retrieval (passed via loaders). Re-enabled them but commented the calling line (it can be re-enabled if I ever need to test the Kong API again).